+++ src/frontend/apps/impress/src/features/left-panel/components/LeftPanelContent.tsx
@@ -1,0 +1,20 @@
+import { useRouter } from 'next/router';
+
+import { Box, SeparatedSection } from '@/components';
+
+import { LeftPanelTargetFilters } from './LefPanelTargetFilters';
+
+export const LeftPanelContent = () => {
+  const router = useRouter();
+  const isHome = router.pathname === '/';
+
+  return (
+    <Box>
+      {isHome && (
+        <SeparatedSection>
+          <LeftPanelTargetFilters />
+        </SeparatedSection>
+      )}
+    </Box>
+  );
+};